globalconnect024/requirements.txt
asgiref==3.8.1
Django==5.2.3
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
psycopg2==2.9.10
PyJWT==2.9.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2

# Required for authentication features
dj-rest-auth==4.0.1
django-allauth==0.61.1
requests==2.31.0
